projects: ad463x: Migrate to new project structure
To support more platforms in the future, migrate the project
to the new multi-platform structure.
