+++ projects/ad463x_fmcz/Makefile
@@ -1,7 +1,13 @@
-IIOD=y
-# Choose the device
+# Uncomment to use the desired platform
+# PLATFORM = xilinx
+# PLATFORM = stm32
+
+# Select the example you want to enable by choosing y for enabling and n for disabling
+BASIC_EXAMPLE = n
+IIO_EXAMPLE = y
+
 # AD4630 Family = 0 | AD4030 Family = 1 | ADAQ4224 = 2
-AD463X_ID=2
+AD463X_ID=0
 
 include ../../tools/scripts/generic_variables.mk
