+++ projects/ad463x_fmcz/src.mk
@@ -1,58 +1,37 @@
-################################################################################
-#									       #
-#     Shared variables:							       #
-#	- PROJECT							       #
-#	- DRIVERS							       #
-#	- INCLUDE							       #
-#	- PLATFORM_DRIVERS						       #
-#	- NO-OS								       #
-#									       #
-################################################################################
+include $(PROJECT)/src/platform/$(PLATFORM)/platform_src.mk
+include $(PROJECT)/src/examples/examples_src.mk
 
-SRCS := $(PROJECT)/src/ad463x_fmc.c
+SRCS += $(PROJECT)/src/platform/$(PLATFORM)/main.c \
+	$(PROJECT)/src/common/common_data.c \
+	$(PROJECT)/src/platform/$(PLATFORM)/parameters.c
+
+INCS += $(PROJECT)/src/common/common_data.h \
+	$(PROJECT)/src/platform/platform_includes.h \
+	$(PROJECT)/src/platform/$(PLATFORM)/parameters.h
+
 SRCS += $(DRIVERS)/api/no_os_spi.c \
-	$(DRIVERS)/api/no_os_pwm.c \
+	$(DRIVERS)/api/no_os_dma.c \
 	$(DRIVERS)/api/no_os_gpio.c \
-	$(DRIVERS)/api/no_os_irq.c \
 	$(DRIVERS)/api/no_os_uart.c \
-	$(DRIVERS)/adc/ad463x/ad463x.c \
-	$(DRIVERS)/axi_core/axi_dmac/axi_dmac.c \
-	$(DRIVERS)/axi_core/clk_axi_clkgen/clk_axi_clkgen.c \
-	$(DRIVERS)/axi_core/axi_pwmgen/axi_pwm.c \
-	$(DRIVERS)/axi_core/spi_engine/spi_engine.c \
-	$(NO-OS)/util/no_os_util.c \
-	$(NO-OS)/util/no_os_alloc.c \
-	$(NO-OS)/util/no_os_mutex.c \
-	$(PLATFORM_DRIVERS)/$(PLATFORM)_uart.c \
-	$(NO-OS)/util/no_os_lf256fifo.c \
-	$(PLATFORM_DRIVERS)/$(PLATFORM)_irq.c \
-	$(NO-OS)/util/no_os_fifo.c \
-	$(NO-OS)/util/no_os_list.c \
-	$(DRIVERS)/api/no_os_irq.c 
+	$(DRIVERS)/api/no_os_pwm.c \
+	$(DRIVERS)/api/no_os_irq.c
 
-SRCS +=	$(PLATFORM_DRIVERS)/xilinx_axi_io.c \
-	$(PLATFORM_DRIVERS)/xilinx_gpio.c \
-	$(PLATFORM_DRIVERS)/xilinx_spi.c \
-	$(PLATFORM_DRIVERS)/xilinx_delay.c
-ifeq (y,$(strip $(IIOD)))
-LIBRARIES += iio
-SRC_DIRS += $(NO-OS)/iio/iio_app
-SRCS += $(DRIVERS)/adc/ad463x/iio_ad463x.c 
-endif
-INCS += $(PROJECT)/src/parameters.h
-INCS += $(DRIVERS)/adc/ad463x/ad463x.h \
-	$(DRIVERS)/axi_core/axi_dmac/axi_dmac.h \
-	$(DRIVERS)/axi_core/clk_axi_clkgen/clk_axi_clkgen.h \
-	$(DRIVERS)/axi_core/axi_pwmgen/axi_pwm_extra.h \
-	$(DRIVERS)/axi_core/spi_engine/spi_engine.h \
-	$(DRIVERS)/axi_core/spi_engine/spi_engine_private.h
 INCS +=	$(PLATFORM_DRIVERS)/$(PLATFORM)_spi.h \
 	$(PLATFORM_DRIVERS)/$(PLATFORM)_irq.h \
 	$(PLATFORM_DRIVERS)/$(PLATFORM)_uart.h \
 	$(PLATFORM_DRIVERS)/$(PLATFORM)_gpio.h
+
+SRCS += $(NO-OS)/util/no_os_util.c \
+	$(NO-OS)/util/no_os_alloc.c \
+	$(NO-OS)/util/no_os_mutex.c \
+	$(NO-OS)/util/no_os_circular_buffer.c \
+	$(NO-OS)/util/no_os_list.c \
+	$(NO-OS)/util/no_os_lf256fifo.c \
+	$(NO-OS)/util/no_os_fifo.c
+
 INCS +=	$(INCLUDE)/no_os_axi_io.h \
 	$(INCLUDE)/no_os_spi.h \
-	$(INCLUDE)/no_os_pwm.h \
+	$(INCLUDE)/no_os_dma.h \
 	$(INCLUDE)/no_os_gpio.h \
 	$(INCLUDE)/no_os_error.h \
 	$(INCLUDE)/no_os_delay.h \
@@ -61,14 +40,16 @@ INCS +=	$(INCLUDE)/no_os_axi_io.h \
 	$(INCLUDE)/no_os_lf256fifo.h \
 	$(INCLUDE)/no_os_util.h \
 	$(INCLUDE)/no_os_units.h \
-	$(INCLUDE)/no_os_print_log.h \
 	$(INCLUDE)/no_os_alloc.h \
-	$(INCLUDE)/no_os_fifo.h \
+	$(INCLUDE)/no_os_mutex.h \
+	$(INCLUDE)/no_os_pwm.h \
+	$(INCLUDE)/no_os_circular_buffer.h \
+	$(INCLUDE)/no_os_print_log.h \
 	$(INCLUDE)/no_os_list.h \
-	$(INCLUDE)/no_os_mutex.h
-ifeq (y,$(strip $(IIOD)))
-INCS += $(DRIVERS)/adc/ad463x/iio_ad463x.h
-endif
+	$(INCLUDE)/no_os_lf256fifo.h \
+	$(INCLUDE)/no_os_irq.h \
+	$(INCLUDE)/no_os_fifo.h
+
 ifeq (2,$(strip $(AD463X_ID)))
 CFLAGS += -DADAQ4224_DEV
 else ifeq (1,$(strip $(AD463X_ID)))
